chore: fix links, fix typo in filename, update documentation structure
