+++ CHANGELOG.md
@@ -7,15 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 The Amazon Web Services (AWS) ODBC Driver for PostgreSQL allows an application to take advantage of additional authentication features such Okta Federated Service and AWS Secret Manager.
 
 ### :magic_wand: Added
-- [AWS Secret Manager authentication support](./docs/authentication.md#secret-manager-authentication)
-- [AWS Federated authentication method with OKTA](./docs/authentication.md#okta-authentication)
+- [AWS Secret Manager authentication support](docs/using-the-aws-driver/authentication/authentication.md#secret-manager-authentication)
+- [AWS Federated authentication method with OKTA](docs/using-the-aws-driver/authentication/authentication.md#okta-authentication)
 
 ## [0.0.1] - 2024-01-30
 The Amazon Web Services (AWS) ODBC Driver for PostgreSQL allows an application to take advantage of the AWS authentication features.
 
 ### :magic_wand: Added
-- The [AWS IAM Authentication method](./docs/authentication.md#iam-authentication)
-- The [AWS Federated authentication method for ADFS](./docs/authentication.md#adfs-authentication)
+- The [AWS IAM Authentication method](docs/using-the-aws-driver/authentication/authentication.md#iam-authentication)
+- The [AWS Federated authentication method for ADFS](docs/using-the-aws-driver/authentication/authentication.md#adfs-authentication)
 
 [0.0.2]: https://github.com/awslabs/aws-pgsql-odbc/releases/tag/0.0.2
 [0.0.1]: https://github.com/awslabs/aws-pgsql-odbc/releases/tag/0.0.1
