+++ README.md
@@ -11,7 +11,7 @@
   - [About the Driver](#about-the-driver)
     - [Benefits of the AWS ODBC Driver for PostgreSQL](#benefits-of-the-aws-odbc-driver-for-postgresql)
   - [Getting Started](#getting-started)
-  - [Using the Driver](#using-the-driver)
+  - [Using the Driver](#build-and-test-the-driver)
   - [Documentation](#documentation)
   - [Getting Help and Opening Issues](#getting-help-and-opening-issues)
   - [License](#license)
@@ -24,11 +24,11 @@ The official open-source [PostgreSQL ODBC driver](https://git.postgresql.org/git
 
 ## Getting Started
 
-For more information on how to start using the AWS ODBC Driver for PostgreSQL with AWS authentication, please visit the [authentication documentation](./docs/authentication.md).
+For more information on how to start using the AWS ODBC Driver for PostgreSQL, please visit the [Getting Started page](./docs/getting_started.md).
 
-## Using the Driver
+## Build and Test the Driver
 
-Please refer to the AWS Driver's [documentation](./docs/Documentation.md) for details on how to use, build, and test the AWS ODBC Driver for PostgreSQL.
+Please refer to the AWS Driver's [documentation](./docs/documentation) for details on how to use, build, and test the AWS ODBC Driver for PostgreSQL.
 
 ## Documentation
 
@@ -36,12 +36,12 @@ Technical documentation regarding the functionality of the AWS ODBC Driver for P
 
 ## Getting Help and Opening Issues
 
-If you encounter an issue or bug with the AWS ODBC Driver for PostgreSQL, we would like to hear about it. Please search the [existing issues](https://github.com/awslabs/aws-mysql-odbc/issues) and see if others are also experiencing the same problem before opening a new issue. When opening a new issue, please make sure to provide:
+If you encounter an issue or bug with the AWS ODBC Driver for PostgreSQL, we would like to hear about it. Please search the [existing issues](https://github.com/aws/aws-pgsql-odbc/issues) and see if others are also experiencing the same problem before opening a new issue. When opening a new issue, please make sure to provide:
 
 - the version of the AWS ODBC Driver for PostgreSQL
 - the OS platform and version
 - the PostgreSQL database version you are running
-  Please include a reproduction case for the issue when appropriate. Also please [include driver logs](./docs/using-the-aws-driver/UsingTheAwsDriver.md#logging) if possible, as they help us diagnose problems quicker.
+  Please include a reproduction case for the issue when appropriate. Also please [include driver logs](./docs/using-the-aws-driver/using_the_aws_driver.md#logging) if possible, as they help us diagnose problems quicker.
 
 The GitHub issues are intended for bug reports and feature requests. Keeping the list of open issues lean will help us respond in a timely manner.
