+++ docs/using-the-aws-driver/authentication/img/database.png
