+++ docs/using-the-aws-driver/authentication/img/secret_manager.png
