+++ docs/using-the-aws-driver/limitless/img/connection_pool.png
